Simplify cors behavior attribute
