--- src/Moryx.Runtime.Wcf/CustomContractBehaviorAttribute.cs
@@ -1,68 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.ServiceModel;
-using System.ServiceModel.Channels;
-using System.ServiceModel.Configuration;
-using System.ServiceModel.Description;
-using System.ServiceModel.Dispatcher;
-
-namespace Moryx.Runtime.Wcf
-{
-    internal class CustomHeaderMessageInspector : IDispatchMessageInspector
-    {
-        readonly Dictionary<string, string> _requiredHeaders;
-
-        public CustomHeaderMessageInspector(Dictionary<string, string> headers)
-        {
-            _requiredHeaders = headers ?? new Dictionary<string, string>();
-        }
-
-        public object AfterReceiveRequest(ref Message request, IClientChannel channel, InstanceContext instanceContext)
-        {
-            return null;
-        }
-
-        public void BeforeSendReply(ref Message reply, object correlationState)
-        {
-            var httpHeader = reply.Properties["httpResponse"] as HttpResponseMessageProperty;
-            foreach (var item in _requiredHeaders)
-            {
-                httpHeader.Headers.Add(item.Key, item.Value);
-            }
-        }
-    }
-
-    internal class CustomContractBehaviorAttribute : BehaviorExtensionElement, IEndpointBehavior
-    {
-        public override Type BehaviorType => typeof(CustomContractBehaviorAttribute);
-
-        public void AddBindingParameters(ServiceEndpoint endpoint, BindingParameterCollection bindingParameters)
-        {
-        }
-
-        public void ApplyClientBehavior(ServiceEndpoint endpoint, ClientRuntime clientRuntime)
-        {
-        }
-
-        public void ApplyDispatchBehavior(ServiceEndpoint endpoint, EndpointDispatcher endpointDispatcher)
-        {
-            var requiredHeaders = new Dictionary<string, string>
-            {
-                {"Access-Control-Allow-Origin", "*"},
-                {"Access-Control-Request-Method", "POST,GET,PUT,DELETE,OPTIONS"},
-                {"Access-Control-Allow-Headers", "X-Requested-With,Content-Type"}
-            };
-
-            endpointDispatcher.DispatchRuntime.MessageInspectors.Add(new CustomHeaderMessageInspector(requiredHeaders));
-        }
-
-        public void Validate(ServiceEndpoint endpoint)
-        {
-        }
-
-        protected override object CreateBehavior()
-        {
-            return new CustomContractBehaviorAttribute();
-        }
-    }
-}
